pyproject.toml
[project]
name = "ad-xray"
version = "0.1.0"
description = "Meta Ads Spy Backend MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "celery>=5.3.6",
    "redis>=5.0.3",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.2.1",
    "playwright>=1.42.0",
    "httpx>=0.27.0",
    "boto3>=1.34.0",
    "yt-dlp>=2024.3.10",
    "google-generativeai>=0.4.1",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "psycopg2-binary>=2.9.9",  # For sync operations if needed, or stick to asyncpg
    "sqlalchemy>=2.0.28",
    "asgiref>=3.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
testpaths = ["tests"]
